Optimize mock and router
