--- ISSUE_TEMPLATE.md
@@ -1,20 +1,0 @@
-Before opening an issue:
-
-- [Search for duplicate or closed issues](https://github.com/coreui/coreui-free-vue-admin-template/issues?utf8=%E2%9C%93&q=is%3Aissue)
-- Prepare a [reduced test case](https://css-tricks.com/reduced-test-cases/) for any bugs
-- Read the [contributing guidelines](https://github.com/coreui/coreui-free-vue-admin-template/blob/master/CONTRIBUTING.md)
-
-When asking general "how to" questions:
-
-- Please do not open an issue here
-
-When reporting a bug, include:
-
-- Operating system and version (Windows, Mac OS X, Android, iOS, Win10 Mobile)
-- Browser and version (Chrome, Firefox, Safari, IE, MS Edge, Opera 15+, Android Browser)
-- Reduced test cases and potential fixes using [CodePen](https://codepen.io/) or [JS Bin](https://jsbin.com/)
-
-When suggesting a feature, include:
-
-- As much detail as possible for what we should add and why it's important to CoreUI Admin Template
-- Relevant links to prior art, screenshots, or live demos whenever possible
