+++ src/mock/doctor.js
@@ -1,0 +1,1 @@
+export default function () {}
