+++ src/mock/tech.js
@@ -1,0 +1,1 @@
+export default function () {}
