+++ src/router/common.js
@@ -1,0 +1,36 @@
+const Page404 = () => import('@/views/pages/Page404')
+const Page500 = () => import('@/views/pages/Page500')
+const Login = () => import('@/views/pages/Login')
+const Register = () => import('@/views/pages/Register')
+
+
+export default {
+  path: '/',
+    redirect: '/login',
+  name: 'Pages',
+  component: {
+  render (c) { return c('router-view') }
+},
+  children: [
+    {
+      path: '404',
+      name: 'Page404',
+      component: Page404
+    },
+    {
+      path: '500',
+      name: 'Page500',
+      component: Page500
+    },
+    {
+      path: 'login',
+      name: 'Login',
+      component: Login
+    },
+    {
+      path: 'register',
+      name: 'Register',
+      component: Register
+    }
+  ]
+}
