+++ src/router/register.js
@@ -1,0 +1,1 @@
+export default {}
